requirements.txt
Flask==2.0.2
redis==3.5.3
mysql-connector-python==8.0.23
requests==2.26.0